Refactoring, renaming, comments
