+++ SpotUp/Models/PlaceList.swift
@@ -11,7 +11,7 @@ import Foundation
 struct PlaceList: Identifiable {
     var id: String = UUID().uuidString
     var name: String
-    var owner: SimpleUser
+    var owner: ListOwner
     var followerIds: [String] = []
     var isPublic: Bool = true
     var placeIds: [String] = []
