+++ SpotUp/Views/ProfileView/ProfileRows.swift
@@ -1,0 +1,48 @@
+//
+//  ProfileRows.swift
+//  SpotUp
+//
+//  Created by Timo Erdelt on 21.11.19.
+//  Copyright © 2019 iOS WiSe 19/20 Gruppe 7. All rights reserved.
+//
+
+import SwiftUI
+
+struct CreateNewPlaceListRow: View {
+    @Binding var showSheet: Bool
+    @Binding var sheetSelection: String
+    
+    var body: some View {
+        Button(action: {
+            print("buttonPressed")
+            self.sheetSelection = "create_placelist"
+            self.showSheet.toggle()
+        }) {
+            HStack {
+                Image(systemName: "plus")
+                    .resizable()
+                    .scaledToFit()
+                    .frame(width: 50, height: 50)
+                Text("Create new place list")
+            }
+        }
+    }
+}
+
+struct PlacesListRow: View {
+    var placeList: PlaceList
+    
+    var body: some View {
+        HStack {
+            Image("chincoteague")
+                .resizable()
+                .scaledToFit()
+                .frame(width: 50, height: 50)
+            VStack(alignment: .leading) {
+                Text(placeList.name)
+                    .bold()
+                Text("von \(placeList.owner.username)")
+            }
+        }
+    }
+}
